pyproject.toml
[project]
name = "seikyu-sho-sender"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.1",
]
